[PageBlob][DownloadSparseBlob]Addressing comments
