+++ sdk/storage/azure-storage-blob/azure/storage/blob/_shared/parser.py
@@ -18,10 +18,3 @@ else:
 
 def _to_utc_datetime(value):
     return value.strftime('%Y-%m-%dT%H:%M:%SZ')
-
-
-def get_empty_chunk(chunk_size):
-    empty_chunk = b''
-    for i in range(0, chunk_size):  # pylint:disable=unused-variable
-        empty_chunk += b'\x00'
-    return empty_chunk
