+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/__init__.py
@@ -39,6 +39,7 @@ from .models import (
     BlobPropertiesPaged,
     BlobPrefix
 )
+from .download_async import StorageStreamDownloader
 from .blob_client_async import BlobClient
 from .container_client_async import ContainerClient
 from .blob_service_client_async import BlobServiceClient
@@ -81,4 +82,5 @@ __all__ = [
     'BlobPermissions',
     'ResourceTypes',
     'AccountPermissions',
+    'StorageStreamDownloader',
 ]
